+++ rana_qgis_plugin/loader.py
@@ -1,8 +1,8 @@
 import os
 from copy import deepcopy
-from datetime import datetime
 from functools import partial
 from pathlib import Path
+from typing import Optional
 
 from qgis.core import QgsDataSourceUri, QgsProject, QgsRasterLayer, QgsSettings
 from qgis.PyQt.QtCore import (
@@ -64,13 +64,9 @@ from rana_qgis_plugin.utils_api import (
     start_tenant_process,
 )
 from rana_qgis_plugin.utils_qgis import (
-    COLOR_RAMP_OCEAN_HALINE,
-    apply_gradient_ramp,
-    color_ramp_from_data,
     convert_vectorfile_to_geopackage,
     get_threedi_results_analysis_tool_instance,
     is_loaded_in_schematisation_editor,
-    multiband_raster_min_max,
 )
 from rana_qgis_plugin.utils_settings import hcc_working_dir
 from rana_qgis_plugin.widgets.result_browser import ResultBrowser
@@ -82,8 +78,11 @@ from rana_qgis_plugin.workers import (
     FileUploadWorker,
     LizardResultDownloadWorker,
     RasterStyleWorker,
+    ProjectJobMonitorWorker,
     VectorStyleWorker,
 )
+
+STYLE_DIR = Path(__file__).parent / "styles"
 
 
 class Loader(QObject):
@@ -117,6 +116,8 @@ class Loader(QObject):
     model_created = pyqtSignal()
     revision_saved = pyqtSignal()
     model_deleted = pyqtSignal()
+    project_jobs_added = pyqtSignal(list)
+    project_job_updated = pyqtSignal(dict)
 
     def __init__(self, communication, parent):
         super().__init__(parent)
@@ -125,6 +126,7 @@ class Loader(QObject):
         self.vector_style_worker: QThread = None
         self.raster_style_worker: QThread = None
         self.new_file_upload_worker: QThread = None
+        self.project_job_monitor: QThread = None
         self.communication = communication
 
         # For simulations
@@ -134,6 +136,9 @@ class Loader(QObject):
         # For upload of schematisations
         self.upload_thread_pool = QThreadPool()
         self.upload_thread_pool.setMaxThreadCount(1)
+
+    def __del__(self):
+        self.stop_project_job_monitoring()
 
     @pyqtSlot(dict, dict)
     def open_wms(self, _: dict, file: dict) -> bool:
@@ -208,18 +213,6 @@ class Loader(QObject):
         )
         self.file_download_finished.emit(None)
 
-    def apply_style(self, layer):
-        # Water level styling
-        min_value, max_value = multiband_raster_min_max(layer)
-        color_ramp = color_ramp_from_data(COLOR_RAMP_OCEAN_HALINE)
-        apply_gradient_ramp(
-            layer=layer,
-            color_ramp=color_ramp,
-            min_value=min_value,
-            max_value=max_value,
-            band=1,
-        )
-
     def on_file_download_finished(
         self, project, file, local_file_path: str, from_thread=True
     ):
@@ -257,7 +250,9 @@ class Loader(QObject):
                                 waterdepth_layer = QgsRasterLayer(
                                     waterdepth_path, "max_waterdepth.tif", "gdal"
                                 )
-                                self.apply_style(waterdepth_layer)
+                                waterdepth_layer.loadNamedStyle(
+                                    str(STYLE_DIR / "water_depth.qml")
+                                )
                                 if hasattr(waterdepth_layer.renderer(), "setBand"):
                                     waterdepth_layer.renderer().setBand(1)
                                 waterdepth_layer.setName("max_waterdepth.tif")
@@ -415,27 +410,18 @@ class Loader(QObject):
         else:
             self.communication.show_warn(f"Unable to create folder {folder_name}")
 
-    @pyqtSlot(dict)
-    @pyqtSlot(dict, int)
-    def create_schematisation_revision_3di_model(self, file, revision_id=None):
-        tc = ThreediCalls(get_threedi_api())
+    @pyqtSlot(dict, dict)
+    @pyqtSlot(dict, dict, int)
+    def create_schematisation_revision_3di_model(self, project, file, revision_id=None):
         # Retrieve schematisation info
         schematisation = get_threedi_schematisation(
             self.communication, file["descriptor_id"]
         )
         if not revision_id:
             revision_id = schematisation["latest_revision"]["id"]
-        schematisation_id = schematisation["schematisation"]["id"]
-        try:
-            tc.create_schematisation_revision_3di_model(schematisation_id, revision_id)
-            self.model_created.emit()
-        except ApiException as e:
-            if e.status == 400:
-                QMessageBox.warning(
-                    QApplication.activeWindow(), "Warning", eval(e.body)[0]
-                )
-            else:
-                raise
+        self.start_model_tracker_process(
+            project, schematisation["schematisation"], revision_id
+        )
 
     @pyqtSlot(dict, int)
     def delete_schematisation_revision_3di_model(self, file, revision_id):
@@ -479,6 +465,13 @@ class Loader(QObject):
         organisations = {
             org.unique_id: org for org in tc.fetch_organisations(allowed_org_ids)
         }
+
+        if len(organisations) == 0:
+            self.communication.show_warn(
+                "No organisation available for this simulation"
+            )
+            self.simulation_started_failed.emit()
+            return
 
         # Retrieve schematisation info
         schematisation = get_threedi_schematisation(
@@ -568,7 +561,7 @@ class Loader(QObject):
                     lizard_post_processing_overview,
                 )
             simulation_wizard.simulation_created.connect(
-                partial(self.start_process, project, file)
+                partial(self.start_simulation_tracker_process, project, file)
             )
             simulation_wizard.simulation_created_failed.connect(
                 self.simulation_started_failed
@@ -577,14 +570,39 @@ class Loader(QObject):
             if simulation_wizard.exec() == QDialog.Rejected:
                 self.simulation_cancelled.emit()
 
-    def start_process(self, project, file, simulations):
-        # Find the simulation tracker processes
+    def get_process_id_for_tag(self, tag: str) -> Optional[str]:
         processes = get_tenant_processes(self.communication)
-        track_process = None
         for process in processes:
-            if "simulation_tracker" in process["tags"]:
-                track_process = process["id"]
-                break
+            if tag in process["tags"]:
+                return process["id"]
+
+    def start_model_tracker_process(
+        self,
+        project,
+        schematisation,
+        revision_id: int,
+        inherit_from_previous_revision: bool = True,
+    ):
+        track_process = self.get_process_id_for_tag("model_tracker")
+        if track_process is None:
+            self.communication.log_err("No model tracker available")
+            return
+        params = {
+            "project_id": project["id"],
+            "inputs": {
+                "schematisation_id": schematisation["id"],
+                "revision_id": revision_id,
+                "inherit_from_previous_model": True,
+                "inherit_from_previous_revision": inherit_from_previous_revision,
+            },
+            "name": f"model_tracker_{schematisation['name']}_rev{revision_id}",
+        }
+        _ = start_tenant_process(self.communication, track_process, params)
+        self.model_created.emit()
+
+    def start_simulation_tracker_process(self, project, file, simulations):
+        # Find the simulation tracker processes
+        track_process = self.get_process_id_for_tag("simulation_tracker")
 
         if track_process is None:
             self.communication.log_err("No simulation tracker available")
@@ -1068,7 +1086,7 @@ class Loader(QObject):
         self.schematisation_upload_finished.emit()
         load_local_schematisation(
             communication=self.communication,
-            local_schematisation=local_schematisation,
+            local_schematisation=local_schematisation.wip_revision,
             action=BuildOptionActions.CREATED,
         )
 
@@ -1169,7 +1187,15 @@ class Loader(QObject):
                 local_schematisation,
                 new_upload,
             )
-
+            upload_worker.signals.create_model_requested.connect(
+                lambda revision_id,
+                inherit_from_previous_revision: self.start_model_tracker_process(
+                    project,
+                    schematisation.to_dict(),
+                    revision_id,
+                    inherit_from_previous_revision,
+                )
+            )
             upload_worker.signals.thread_finished.connect(
                 self.on_schematisation_upload_finished
             )
@@ -1268,3 +1294,17 @@ class Loader(QObject):
 
         self.upload_thread_pool.start(upload_worker)
         self.revision_saved.emit()
+
+    def stop_project_job_monitoring(self):
+        if self.project_job_monitor:
+            self.project_job_monitor.stop()
+
+    def start_project_job_monitoring(self, project_id):
+        self.stop_project_job_monitoring()
+        self.project_job_monitor = ProjectJobMonitorWorker(
+            project_id=project_id, parent=self
+        )
+        self.project_job_monitor.jobs_added.connect(self.project_jobs_added)
+        self.project_job_monitor.job_updated.connect(self.project_job_updated)
+        self.project_job_monitor.failed.connect(self.communication.show_warn)
+        self.project_job_monitor.start()